Max Score is 10

+++ src/view/Student/Score.java
@@ -8,7 +8,7 @@ import java.awt.event.*;
 
 public class Score extends JFrame implements ActionListener {
 
-    Score(int score) {
+    Score(double score) {
         setBounds(400, 150, 750, 550);
         getContentPane().setBackground(Color.WHITE);
         setLayout(null);
